refactor: split coordination.rs into coordination/ module dir (pure move)
coordination.rs -> coordination/checkpoint.rs (workflow-checkpoint kind +
its tests, verbatim) with the shared protocol core (§9 rules doc, error
type, grammar, space key, defensive open, modes, ulid) extracted to
coordination/mod.rs. Only non-move lines: the mod declaration/re-export
in mod.rs and the use-super glue in checkpoint.rs. No logic changes.
